ai/requirements.txt
fastapi
uvicorn
pydantic
openai
langchain
faiss-cpu
python-dotenv
requests
